move:#4 서버 액션 및 상수 폴더 이동
