+++ src/app/category/top-rated/[page]/page.tsx
@@ -1,6 +1,6 @@
 import PageNationBtn from '@/components/pageNationBtn';
 import { Movie } from '@/types/category/movie';
-import { getTopRated } from '@/utils/tmdb/serverApi';
+import { getTopRated } from '@/services/category/serverApi';
 import React from 'react';
 
 // 서버 컴포넌트에서 쿼리 파라미터 가져오기
